Perfect conductor boundary conditions (top & bottom)
This implementation has been proposed by Domenico Meduri and Laurene Jouve.

It is similar to the implementation that is currently used in the ASH code.

To deal with the over-constraint on the poloidal component, the condition

(1) Br=0 ==> B_p = 0

is applied at the boundaries n_r=(1,n_r_max) and the condition on the current

(2) n x j= 0

is applied in the domain n_r=(2,n_r_max-1).

In some cases, this compromise can be more stable than just relaxing
the constraint (1) which is supposed to be sage only if
the initial conditions do match the boundary conditions.

See Glatzmaier, 1984, JCP, 55, 461

" [...] as pointed out by P. H. Roberts (private communication),
only the second [...] of these are required since the first automatically
vanishes at the bottom when u, and B, vanish there due to the radial
component of the induction equation assuming Blm was initially zero [...] "